Move app component to match desired directory structure
