--- src/alphanum-sort.d.ts
@@ -1,1 +1,0 @@
-declare module "alphanum-sort";
