+++ src/index.jsx
@@ -3,7 +3,7 @@ import { Provider } from 'react-redux';
 import { render } from 'react-dom';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import configureStore from './store/configureStore';
-import App from './App';
+import App from './components/App/App';
 import './index.css';
 
 const store = configureStore();
